Move ControllerMap to proxy pkg and other nits
